backend/requirements.txt
annotated-types==0.6.0
asgiref==3.6.0
attrs==23.2.0
blis==0.7.11
branca==0.7.0
catalogue==2.0.10
certifi==2023.11.17
charset-normalizer==3.3.2
click==8.1.7
click-plugins==1.1.1
cligj==0.7.2
cloudpathlib==0.16.0
confection==0.1.4
contourpy==1.2.0
cycler==0.12.1
cymem==2.0.8
Django==4.1.7
django-cors-headers==4.3.1
python-dotenv
django-crispy-forms==2.0
en-core-web-sm @ https://github.com/explosion/spacy-models/releases/download/en_core_web_sm-3.7.1/en_core_web_sm-3.7.1-py3-none-any.whl#sha256=86cc141f63942d4b2c5fcee06630fd6f904788d2f0ab005cce45aadb8fb73889
et-xmlfile==1.1.0
fastapi
folium==0.15.1
fonttools==4.47.0
geographiclib==2.0
geopy==2.4.1
idna==3.6
Jinja2==3.1.3
joblib==1.3.2
kiwisolver==1.4.5
langcodes==3.3.0
MarkupSafe==2.1.3
matplotlib==3.8.2
murmurhash==1.0.10
networkx==3.2.1
nltk==3.8.1
numpy==1.26.3
openpyxl==3.1.1

packaging==23.2
pandas==1.5.3
pillow==10.2.0
preshed==3.0.9
psycopg2-binary
pydantic==2.6.1
pydantic_core==2.16.2
pyparsing==3.1.1
pyproj==3.6.1
python-dateutil==2.8.2
pytz==2022.7.1
regex==2023.12.25
requests==2.31.0
scikit-learn==1.3.2
scipy==1.11.4
setuptools==69.0.3
shapely==2.0.2
six==1.16.0
smart-open==6.4.0
spacy==3.7.2
spacy-legacy==3.0.12
spacy-loggers==1.0.5
sqlparse==0.4.3
sqlalchemy
srsly==2.4.8
thinc==8.2.3
threadpoolctl==3.2.0
tqdm==4.66.1
typer==0.9.0
typing_extensions==4.9.0
tzdata==2023.4
urllib3==2.1.0
uvicorn
wasabi==1.1.2
weasel==0.3.4
wordcloud==1.9.3
XlsxWriter==3.0.8
passlib
pydantic[email]
bcrypt
python-multipart
